+++ src/Jobs/VimeoUpload.php
@@ -37,28 +37,34 @@ class VimeoUpload implements ShouldQueue
      */
     public function handle()
     {
-        Log::info("ThotamFileLibrary upload to Vimeo: ".$this->fileUpload->id. " - starting");
+        $this->fileUpload = FileLibrary::find($this->fileUpload->id);
 
-        $parameters = [
-            'name' => $this->fileUpload->vimeo_name,
-            'description' => $this->fileUpload->vimeo_description,
-            'privacy' => [
-                'view' => $this->fileUpload->vimeo_view,
-            ],
-        ];
+        if (!!$this->fileUpload->vimeo_id) {
+            Log::info("ThotamFileLibrary upload to Vimeo: ".$this->fileUpload->id. " - uploaded");
+        } else {
+            Log::info("ThotamFileLibrary upload to Vimeo: ".$this->fileUpload->id. " - starting");
 
-        $response = Vimeo::upload(Storage::disk('public')->path($this->fileUpload->local_path), $parameters);
+            $parameters = [
+                'name' => $this->fileUpload->vimeo_name,
+                'description' => $this->fileUpload->vimeo_description,
+                'privacy' => [
+                    'view' => $this->fileUpload->vimeo_view,
+                ],
+            ];
 
-        $this->fileUpload->update([
-            "vimeo_id" => Str::of($response)->explode('/')->last(),
-        ]);
+            $response = Vimeo::upload(Storage::disk('public')->path($this->fileUpload->local_path), $parameters);
 
-        $check = FileLibrary::find($this->fileUpload->id);
-        //if (!!$check->google_id && !!$check->youtube_id) {
-        if (!!$check->google_id) {
-            Storage::disk('public')->delete($this->fileUpload->local_path);
+            $this->fileUpload->update([
+                "vimeo_id" => Str::of($response)->explode('/')->last(),
+            ]);
+
+            $check = FileLibrary::find($this->fileUpload->id);
+            //if (!!$check->google_id && !!$check->youtube_id) {
+            if (!!$check->google_id) {
+                Storage::disk('public')->delete($this->fileUpload->local_path);
+            }
+
+            Log::info("ThotamFileLibrary upload to Vimeo: ".$this->fileUpload->id. " - finished");
         }
-
-        Log::info("ThotamFileLibrary upload to Vimeo: ".$this->fileUpload->id. " - finished");
     }
 }
