+++ src/Jobs/YoutubeUpload.php
@@ -36,22 +36,28 @@ class YoutubeUpload implements ShouldQueue
      */
     public function handle()
     {
-        Log::info("ThotamFileLibrary upload to Youtube: ".$this->fileUpload->id. " - starting");
+        $this->fileUpload = FileLibrary::find($this->fileUpload->id);
 
-        $youtube = new Youtube;
+        if (!!$this->fileUpload->youtube_id) {
+            Log::info("ThotamFileLibrary upload to Youtube: ".$this->fileUpload->id. " - uploaded");
+        } else {
+            Log::info("ThotamFileLibrary upload to Youtube: ".$this->fileUpload->id. " - starting");
 
-        $response = $youtube->upload(Storage::disk('public')->path($this->fileUpload->local_path), $this->fileUpload->youtube_data, $this->fileUpload->youtube_privacy_status);
+            $youtube = new Youtube;
 
-        $this->fileUpload->update([
-            "youtube_id" => $response->videoId,
-        ]);
+            $response = $youtube->upload(Storage::disk('public')->path($this->fileUpload->local_path), $this->fileUpload->youtube_data, $this->fileUpload->youtube_privacy_status);
 
-        $check = FileLibrary::find($this->fileUpload->id);
-        //if (!!$check->google_id && !!$check->vimeo_id) {
-        if (!!$check->google_id) {
-            Storage::disk('public')->delete($this->fileUpload->local_path);
+            $this->fileUpload->update([
+                "youtube_id" => $response->videoId,
+            ]);
+
+            $check = FileLibrary::find($this->fileUpload->id);
+            //if (!!$check->google_id && !!$check->vimeo_id) {
+            if (!!$check->google_id) {
+                Storage::disk('public')->delete($this->fileUpload->local_path);
+            }
+
+            Log::info("ThotamFileLibrary upload to Youtube: ".$this->fileUpload->id. " - finished");
         }
-
-        Log::info("ThotamFileLibrary upload to Youtube: ".$this->fileUpload->id. " - finished");
     }
 }
